+++ frontend/package.json
@@ -18,6 +18,7 @@
     "@tanstack/react-query": "^5.59.16",
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
+    "fast-deep-equal": "^3.1.3",
     "lodash": "^4.17.21",
     "msw": "^2.6.0",
     "react": "^18.3.1",
